Refatora inicial separando GUI do main
