+++ README.md
@@ -1,1 +1,13 @@
 # Editor-Autom-tico
+
+Este repositório contém um editor de vídeo simplificado. Para iniciar a interface gráfica utilize:
+
+```bash
+python main.py
+```
+
+Para apenas exibir a ajuda sem carregar dependências gráficas execute:
+
+```bash
+python main.py --help
+```
